+++ src/App.tsx
@@ -1,10 +1,9 @@
-import { useTabProps, useTabRef } from './features/Tab';
+import { useTab } from './features/Tab';
 import { Tab } from './features/Tab';
 
 const App = () => {
-  const { select: firstSelect, getTabItemProps } =
-    useTabProps();
-  const { select: secondSelect, tabListRef } = useTabRef(2);
+  const { getTabItemProps, getTabScreen } = useTab();
+  const { tabListRef, select } = useTab(2);
   return (
     <>
       <div>
@@ -18,15 +17,13 @@ const App = () => {
         </div>
         <hr />
         <div>
-          {
-            [
-              <div>Tab Screen 0</div>,
-              <div>Tab Screen 1</div>,
-              <div>Tab Screen 2</div>,
-              <div>Tab Screen 3</div>,
-              <div>Tab Screen 4</div>,
-            ][firstSelect]
-          }
+          {getTabScreen([
+            <div>Tab Screen 0</div>,
+            <div>Tab Screen 1</div>,
+            <div>Tab Screen 2</div>,
+            <div>Tab Screen 3</div>,
+            <div>Tab Screen 4</div>,
+          ])}
         </div>
       </div>
       <div>
@@ -49,12 +46,12 @@ const App = () => {
               <div>Tab Screen 2</div>,
               <div>Tab Screen 3</div>,
               <div>Tab Screen 4</div>,
-            ][secondSelect]
+            ][select]
           }
         </div>
       </div>
       <Tab.Box
-        onChange={index => console.log(index)}
+        onChange={(index) => console.log(index)}
         initial={2}>
         <h1>Tab Component</h1>
         <Tab.Screens>
@@ -63,7 +60,7 @@ const App = () => {
           <Tab.Screen>thumbnail3</Tab.Screen>
         </Tab.Screens>
         <hr />
-        <Tab.List style={{ display: 'flex' }}>
+        <Tab.List style={{ display: 'flex', gap: '10px' }}>
           <Tab>tab1</Tab>
           <Tab>tab2</Tab>
           <Tab>tab3</Tab>
