+++ src/features/Tab/components/TabBox.tsx
@@ -1,18 +1,14 @@
 import React from 'react';
 import { TabOptions } from '../types/tabOptions';
-import { useTabProps } from '..';
+import { useTab } from '..';
 import { TabContext } from '../hooks/useTabContext';
 
 type TabBoxProps = TabOptions & {
   children: React.ReactNode;
 };
 
-const TabBox = ({
-  initial,
-  onChange,
-  children,
-}: TabBoxProps) => {
-  const { select, getTabItemProps } = useTabProps({
+const TabBox = ({ initial, onChange, children }: TabBoxProps) => {
+  const { select, getTabItemProps } = useTab({
     initial,
     onChange,
   });
