+++ src/features/Tab/index.ts
@@ -1,3 +1,2 @@
-export { default as useTabProps } from './hooks/useTabProps';
-export { default as useTabRef } from './hooks/useTabRef';
+export { default as useTab } from './hooks/useTab';
 export { default as Tab } from './components/Tab';
